pyproject.toml
[tool.poetry]
name = "django-pro"
version = "0.1.0"
description = ""
authors = ["lyh <1363998021@qq.com>"]
readme = "README.md"
packages = [{include = "django_pro"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.4"
djangorestframework = "^3.14.0"
python-dotenv = "^1.0.0"
pymysql = "^1.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
